Feature(payments): корректная работа опции "Спрятать копейки"
TICKER: ANDROID-149

+++ app/src/main/java/com/infinity_coder/divcalendar/presentation/_common/delegate/SecurityCurrencyDelegate.kt
@@ -33,7 +33,7 @@ object SecurityCurrencyDelegate {
 
     fun getValueConsiderCopecks(value: BigDecimal): String {
         return if (settingsRepository.isHideCopecks())
-            formatterWithoutPoints.format(value).toString()
+            formatterWithoutPoints.format(value.toBigInteger()).toString()
         else
             formatterWithPoints.format(value).toString()
     }
